Extract viewport lens pass
